REFACTOR: Move parseCss to common and add new report helper
