+++ src/common/getEmptyReport/index.js
@@ -1,0 +1,314 @@
+export function getEmptyReport() {
+	return {
+		styleSheetSize: {
+			source: 0,
+			gzipSource: 0,
+		},
+		comments: {
+			total: 0,
+			length: {
+				total: 0,
+				longest: 0,
+				shortest: Number.MAX_SAFE_INTEGER,
+				average: 0,
+			},
+			sizeRatio: 0,
+		},
+		atRules: {
+			total: 0,
+			empty: 0,
+			prefixed: 0,
+			unknown: {
+				total: 0,
+				unique: 0,
+				usage: {},
+			},
+			usage: {},
+		},
+		mediaQueries: {
+			total: 0,
+			unique: 0,
+			onlyKeyword: 0,
+			types: {
+				total: 0,
+				unique: 0,
+				usage: {},
+			},
+			features: {
+				total: 0,
+				unique: 0,
+				prefixed: 0,
+				usage: {},
+			},
+			usage: {},
+		},
+		keyframes: {
+			stepsChains: {},
+			longestStepsChain: null,
+			longestStepsChainLength: 0,
+			longestStepsChainAnimation: null,
+			shortestStepsChain: null,
+			shortestStepsChainLength: Number.MAX_SAFE_INTEGER,
+			shortestStepsChainAnimation: null,
+			animatableProperties: [],
+		},
+		rules: {
+			total: 0,
+			empty: 0,
+			withoutTrailingSemicolon: 0,
+		},
+		selectors: {
+			total: 0,
+			unique: 0,
+			baseUsage: {},
+			pseudoClassesUsage: {},
+			pseudoElementsUsage: {},
+			combinators: {
+				total: 0,
+				adjacentSibling: 0,
+				child: 0,
+				descendant: 0,
+				generalSibling: 0,
+			},
+			complex: 0,
+			maxPerRule: 0,
+			averagePerRule: 0,
+			length: {
+				total: 0,
+				longest: 0,
+				longestSelector: null,
+				average: 0,
+			},
+			specificity: {
+				total: [0, 0, 0],
+				highest: [0, 0, 0],
+				highestSelector: null,
+				average: [0, 0, 0],
+				graphData: [],
+			},
+			sizeRatio: 0,
+			usage: {},
+		},
+		declarations: {
+			total: 0,
+			unique: 0,
+			uniqueRatio: 0,
+			important: 0,
+			averagePerRule: 0,
+			length: {
+				total: 0,
+				longest: 0,
+				longestDeclaration: null,
+				average: 0,
+			},
+			sizeRatio: 0,
+			inAtRules: {},
+			list: [],
+		},
+		properties: {
+			total: 0,
+			unique: 0,
+			shorthands: 0,
+			prefixed: 0,
+			unitless: 0,
+			resetsViaAll: 0,
+			negativeMargins: 0,
+			anonymousReplacedElements: 0,
+			performanceHacks: {},
+			explicitDefaultingKeywords: {
+				total: 0,
+				unique: 0,
+				usage: {},
+			},
+			engineTriggers: {
+				composite: {
+					blink: 0,
+					edgehtml: 0,
+					gecko: 0,
+					webkit: 0,
+				},
+				layout: {
+					blink: 0,
+					edgehtml: 0,
+					gecko: 0,
+					webkit: 0,
+				},
+				paint: {
+					blink: 0,
+					edgehtml: 0,
+					gecko: 0,
+					webkit: 0,
+				},
+			},
+			usage: {},
+		},
+		displays: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
+		positions: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
+		zIndices: {
+			total: 0,
+			unique: 0,
+			invalid: {},
+			usage: {},
+		},
+		floats: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
+		borderRadiuses: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
+		letterSpacings: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
+		fontSizes: {
+			total: 0,
+			unique: 0,
+			keywords: {
+				total: 0,
+				unique: 0,
+				usage: {},
+			},
+			usage: {},
+		},
+		lineHeights: {
+			total: 0,
+			unique: 0,
+			hardCoded: 0,
+			usage: {},
+		},
+		fontFamilies: {
+			total: 0,
+			unique: 0,
+			generic: {},
+			system: {},
+			withoutFallbackFonts: 0,
+			imageReplacementHacks: 0,
+			usage: {},
+		},
+		colors: {
+			total: 0,
+			unique: 0,
+			currentColorKeyword: 0,
+			transparentKeyword: 0,
+			models: {},
+			named: {},
+			system: {},
+			usage: {},
+		},
+		backgroundColors: {
+			total: 0,
+			unique: 0,
+			currentColorKeyword: 0,
+			transparentKeyword: 0,
+			models: {},
+			named: {},
+			system: {},
+			usage: {},
+		},
+		allColors: {
+			total: 0,
+			unique: 0,
+			currentColorKeyword: 0,
+			transparentKeyword: 0,
+			models: {},
+			named: {},
+			system: {},
+			usage: {},
+		},
+		transitions: {
+			properties: {},
+			longestDuration: 0,
+			shortestDuration: Number.MAX_SAFE_INTEGER,
+			longestDelay: 0,
+			shortestDelay: Number.MAX_SAFE_INTEGER,
+			timingFunctions: {},
+			invalidTimingFunctions: {},
+		},
+		animations: {
+			total: 0,
+			unique: 0,
+			infinite: 0,
+			longestDuration: 0,
+			shortestDuration: Number.MAX_SAFE_INTEGER,
+			longestDelay: 0,
+			shortestDelay: Number.MAX_SAFE_INTEGER,
+			usage: {},
+			timingFunctions: {},
+			invalidTimingFunctions: {},
+		},
+		functions: {
+			total: 0,
+			unique: 0,
+			prefixed: 0,
+			usage: {},
+		},
+		filters: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
+		gradients: {
+			total: 0,
+			unique: 0,
+			usage: {},
+		},
+		dataUris: {
+			total: 0,
+			unique: 0,
+			length: {
+				total: 0,
+				longest: 0,
+				longestDataUri: null,
+				average: 0,
+			},
+			sizeRatio: 0,
+			usage: {},
+		},
+		units: {
+			total: 0,
+			unique: 0,
+			usage: {},
+			excessive: {
+				total: 0,
+				unique: 0,
+				usage: {},
+			},
+		},
+		variables: {
+			total: 0,
+			unique: 0,
+			usage: {},
+			valuesMap: {},
+		},
+		vendorPrefixes: {
+			total: 0,
+			unique: 0,
+			unknown: {
+				total: 0,
+				unique: 0,
+				usage: {},
+			},
+			usage: {},
+		},
+		browserHacks: {
+			total: 0,
+			usage: {
+				supports: {},
+				media: {},
+			},
+		},
+	};
+}
