+++ src/handlers/properties/handleFloat/index.spec.js
@@ -1,4 +1,5 @@
-import { parseCss } from '../../../converters/parseCss';
+import { getEmptyReport } from '../../../common/getEmptyReport';
+import { parseCss } from '../../../common/parseCss';
 import { handleFloat } from '.';
 
 describe('Module: handleFloat', () => {
@@ -25,13 +26,7 @@ describe('Module: handleFloat', () => {
 	let report;
 
 	beforeEach(() => {
-		report = {
-			floats: {
-				total: 0,
-				unique: 0,
-				usage: {},
-			},
-		};
+		report = getEmptyReport();
 
 		cssRoot.walkDecls((decl) => {
 			handleFloat(decl, report);
